refactor(tools): convert audiotap from CLI to library
Split main.swift into focused modules:
- AppAudioCapture: CATapDescription + IOProc → FileHandle
- MicCaptureHandler: AVAudioEngine → WAV
- AudioCaptureSession: orchestrator (start/stop, computes micDelay)
- AudioCaptureResult: result struct
- Helpers: machTicksToSeconds, getDefaultOutputDeviceUID, writeAllToFileHandle

Key changes:
- Package.swift: executableTarget → library product (AudioTapLib)
- writeAllToStdout → writeAllToFileHandle (generic file descriptor)
- fputs(stderr) → os.log Logger
- Thread.sleep in device change handler → DispatchQueue.main.asyncAfter
- All public API types/methods marked public
- File handle closed on capture failure (prevents fd leak)
- Re-entrant device change guard (isRestarting flag)

This enables the Mac App Store distribution path by removing the
subprocess dependency. The host app imports AudioTapLib directly.
